backend fixed

+++ backend/src/main/java/com/fincore/BankingManagement/Milestone2/CreditCheck/Repository/LoanApplicationRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface LoanApplicationRepository extends JpaRepository<LoanApplication, Integer> {
+public interface LoanApplicationRepository extends JpaRepository<LoanApplication, Long> {
     Optional<LoanApplication> findTopByCustomerCustomerIdOrderByApplicationDateDesc(Long customerId);
 }
